1_parsing/requirements.txt
llama-index==0.12.25
llama-index-core==0.12.25
llama-index-node-parser-docling==0.3.1
llama-index-readers-docling==0.3.2
pypdf2>=3.0.1
easyocr>=1.7.2
fastapi
uvicorn
python-multipart


#pip install -r requirements.txt --index-url https://download.pytorch.org/whl/cu121